+++ src/test/java/jscover/instrument/NodeTestHelper.java
@@ -1,6 +1,5 @@
 package jscover.instrument;
 
-import org.mozilla.javascript.Node;
 import org.mozilla.javascript.ast.AstNode;
 import org.mozilla.javascript.ast.AstRoot;
 import org.mozilla.javascript.ast.NodeVisitor;